requirements.txt
telepot==12.7
python-dotenv==0.18.0
